+++ src/main.rs
@@ -1,15 +1,11 @@
-use amethyst::core::transform::TransformBundle;
-use amethyst::input::{InputBundle, StringBindings};
-use amethyst::prelude::*;
-use amethyst::renderer::RenderingBundle;
-use amethyst::renderer::plugins::{RenderFlat2D, RenderToWindow};
-use amethyst::renderer::types::DefaultBackend;
 use amethyst::utils::application_root_dir;
+use amethyst::{GameDataBuilder, CoreApplication};
+use amethyst::core::TransformBundle;
+use amethyst::renderer::{RenderingBundle, RenderToWindow, RenderFlat2D, types::DefaultBackend};
 use amethyst::tiles::{RenderTiles2D, MortonEncoder};
+use amethyst::network::simulation::laminar::{LaminarSocket, LaminarNetworkBundle, LaminarConfig};
+use std::time::Duration;
 
-use log::info;
-
-mod state;
 mod systems;
 mod entities;
 mod components;
@@ -21,9 +17,6 @@ mod network;
 #[cfg(test)]
 mod test_helpers;
 
-/// Desert sand color
-const BACKGROUND_COLOR: [f32; 4] = [0.75, 0.65, 0.5, 1.0];
-
 fn main() -> amethyst::Result<()> {
     amethyst::start_logger(Default::default());
 
@@ -31,24 +24,32 @@ fn main() -> amethyst::Result<()> {
     let resources_dir = app_root.join("resources");
     let display_config = resources_dir.join("display_config.ron");
 
+    let laminar_config = {
+        let mut conf = LaminarConfig::default();
+        // send heartbeat in every 3 seconds
+        conf.heartbeat_interval = Some(Duration::from_secs(3));
+        conf
+    };
+    let socket = LaminarSocket::bind_with_config("127.0.0.1:1234", laminar_config).unwrap();
+
     let game_data = GameDataBuilder::default()
         .with_bundle(TransformBundle::new())?
         .with_bundle(RenderingBundle::<DefaultBackend>::new()
             .with_plugin(
                 RenderToWindow::from_config_path(display_config)?
-                    .with_clear(BACKGROUND_COLOR)
+                    .with_clear([0.0, 0.0, 0.0, 1.0])
             )
             .with_plugin(RenderFlat2D::default())
-            .with_plugin(RenderTiles2D::<resources::GroundTile, MortonEncoder>::default())
-        )?;
+            .with_plugin(RenderTiles2D::<resources::GroundTile, MortonEncoder>::default()))?
+        .with_bundle(LaminarNetworkBundle::new(Some(socket)))?;
 
-    let mut game = CoreApplication::<_, events::WestinyEvent, events::WestinyEventReader>::build(
-        resources_dir,
-        states::client_states::ConnectState::default(),
-    )?.build(game_data)?;
+    let mut game =
+        CoreApplication::<_, events::WestinyEvent, events::WestinyEventReader>::build(
+            resources_dir,
+            states::connection::ConnectState::default(),
+        )?.build(game_data)?;
 
-    info!("Starting...");
+    log::info!("Starting client");
     game.run();
-
     Ok(())
 }
