+++ src/states/mod.rs
@@ -1,2 +1,3 @@
-pub mod client_states;
+pub mod connection;
 pub mod server_states;
+mod game_states;
